+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - update rust dependencies (might affect wasm runtime)
 
+- fix wasm runtime diagnostic logging not showing up
+
 ## 0.54.3 (2024-09-20)
 
 - fix mmd wasm instance initialization on browser native module loading